Add files via upload
This is the  Iris flower dataset, consisting  of three species: setosa, versicolor, and
and virginica. These species can be distinguished based on their
measurements. The objective  of the exercise was  to train a machine learning model that can learn from
these measurements and accurately classify the Iris flowers into
their respective species.